check_headers: Check for redundant #includes
This is a rather ugly and crude attempt at detecting redundant
#includes, primarily intended for use on the linux kernel tree.

There are two parts: A hackish perl frontend, check_headers, and a
small C program built on the sparse library, cident.

cident is used to extract all global identifiers from a given source
file. It is mostly a ripoff of the example c2xml from the sparse
library. It's output is meant to be machine-readable, so for now it
simply consists of three tab-separated columns. I'm sure this
functionality already exists in some other program, but I haven't
found it.

check_headers processes the files given on the command line. For each,
it starts by doing some very basic preprocessing (mostly stripping
comments to reduce the chance of false negatives). Then it looks for
#include directives. For each of these, it asks cident to produce a
list of global identifiers (struct tags, functions, extern variables,
typedefs etc.). For efficiency, the results are cached, so each header
is only parsed once. Then check_headers checks whether any of the
identifiers provided by the header is actually used; if not, a message
is printed.

Now, it is certainly possible that foo.c includes a.h, a.h includes
b.h, and foo.c doesn't use anything from a.h but does use something
from b.h. In that case, removing the include of a.h would break the
build, but arguably foo.c should include b.h directly instead. In some
cases, though, a.h acts as a wrapper for b.h (in the kernel case
typically linux/a.h including asm/a.h), and b.h should not be included
directly. check_headers contains some heuristics and hard-coded
exceptions to handle this, but they can certainly be improved.