Pplight: Make it object oriented
I have a few ideas for extra things I'd like Pplight to do, and for
some of these it might be nice if we had an object to hold a little
state (such as 'where did this come from' and 'which
pseudo-preprocessing have been done') instead of just a string. This
oo'ifies Pplight.pm.