check_includes: require include guard
The currently blacklisted headers have one thing in common: They don't
contain an include guard (and many of them are in fact meant to be
included multiple times). Since the vast majority of headers do have
an include guard and since those without tend to be of the audit_*
kind which is not parsable in itself anyway, let's require an include
guard for now.