locstat: measure header bloat
This is a rather trivial script I wrote to get some numbers for the
impact of liberal sprinkling of #includes, which are almost never
removed again. It is specifically targeted at the linux kernel and its
generated .*.o.cmd dependency files.

For each given dependency file, we accumulate a few metrics for the
header files it mentions (size in bytes and two different line counts)
and the similar metrics for the .c file itself.

The output consists of a few tab-separated fields, which can then be
processed by whatever statistics tool you wish.

Usage is simply

  find . -name '.*.o.cmd' | sort | xargs /path/to/locstat

or if you have GNU Parallel installed:

  find . -name '.*.o.cmd' | sort | parallel -k -m /path/to/locstat