get_include_guard: extract include guard from header file
This is a trivial wrapper for the function provided by Pplight to get
the name of the macro used as "include guard" for a given header
file. This can be useful for e.g. finding headers missing such an
include guard, and for finding clashes between the identifiers used
across different headers (typically due to careless copy-pasting).